+++ tests/Silex/Tests/Provider/TwigServiceProviderTest.php
@@ -13,6 +13,7 @@ namespace Silex\Tests\Provider;
 
 use Silex\Application;
 use Silex\Provider\TwigServiceProvider;
+use Silex\Provider\HttpFragmentServiceProvider;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -41,8 +42,13 @@ class TwigServiceProviderTest extends \PHPUnit_Framework_TestCase
 
     public function testRenderFunction()
     {
+        if (!class_exists('Symfony\Component\HttpFoundation\RequestStack')) {
+            $this->markTestSkipped();
+        }
+
         $app = new Application();
 
+        $app->register(new HttpFragmentServiceProvider());
         $app->register(new TwigServiceProvider(), array(
             'twig.templates'    => array(
                 'hello' => '{{ render("/foo") }}',